recon-ia/api/requirements.txt
flask
python-whois
python-nmap
